requirements.txt
# Python dependencies for VibeGuard Auditor

# Core dependencies
PyYAML>=6.0.1

# Development dependencies (optional, for testing)
# pytest>=7.4.0
# pytest-cov>=4.1.0
